Minor updates for Win32.

+++ cutil/lm_ssl.c
@@ -616,7 +616,7 @@ value lm_ssl_bind(value v_info, value v_addr, value v_port)
     /* Get the address */
     sin.sin_family = AF_INET;
     sin.sin_addr.s_addr = *(uint32 *)v_addr;
-    sin.sin_port = htons(Int_val(v_port));
+    sin.sin_port = htons((short) Int_val(v_port));
 
     /* Perform the bind */
     info = SslInfo_val(v_info);
@@ -702,7 +702,7 @@ value lm_ssl_connect(value v_info, value v_addr, value v_port)
     /* Get the address */
     sin.sin_family = AF_INET;
     sin.sin_addr.s_addr = *(uint32 *)v_addr;
-    sin.sin_port = htons(Int_val(v_port));
+    sin.sin_port = htons((short) Int_val(v_port));
 
     /* Make the connection */
     info = SslInfo_val(v_info);
